move GovUKPayment interface to @planx/components/Pay/model
